Adding a lot of logic for game and player, splitting into two packages
