+++ src/com/theironyard/clt/Main.java
@@ -1,8 +1,10 @@
 package com.theironyard.clt;
 
+import com.theironyard.game.*;
+
 public class Main {
 
     public static void main(String[] args) {
-        Game game = new Game();
+        new Game();
     }
 }
